requirements.txt
# Flask core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# AI and ML
openai==1.54.0
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
numpy<2.0.0
tiktoken==0.5.2

# HTTP requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
Werkzeug==3.0.1

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
